Moving Note and adding the beginning of Task
